* win32/win32.c (rb_w32_open_osfhandle): reverted to old definition.   [ ruby-Bugs-16948 ]
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/branches/ruby_1_8@14973 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ win32/win32.c
@@ -1836,7 +1836,7 @@ init_stdhandle(void)
 
 #ifdef __BORLANDC__
 static int
-rb_w32_open_osfhandle(intptr_t osfhandle, int flags)
+rb_w32_open_osfhandle(long osfhandle, int flags)
 {
     int fd = _open_osfhandle(osfhandle, flags);
     if (fd == -1) {
